+++ src/main/java/com/project/openmarket/domain/user/controller/UserController.java
@@ -7,7 +7,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.project.openmarket.domain.auth.entity.enums.Role;
+import com.project.openmarket.global.auth.enums.Role;
+import com.project.openmarket.global.auth.enums.SessionConst;
 import com.project.openmarket.domain.user.dto.reposne.UserResponseDto;
 import com.project.openmarket.domain.user.dto.request.LoginRequestDto;
 import com.project.openmarket.domain.user.service.ConsumerService;
