+++ src/main/java/com/project/openmarket/domain/user/dto/request/LoginRequestDto.java
@@ -1,6 +1,6 @@
 package com.project.openmarket.domain.user.dto.request;
 
-import com.project.openmarket.validator.Validator;
+import com.project.openmarket.global.validator.Validator;
 
 import jakarta.validation.constraints.NotEmpty;
